refactor(create-esmx): optimize test code structure and readability
- Simplify configuration validation using toMatchObject
- Remove redundant assertions and comments
- Maintain code clarity while reducing verbosity
- Follow project testing standards
