+++ packages/create-esmx/package.json
@@ -5,7 +5,7 @@
     "type": "module",
     "private": false,
     "bin": {
-        "create-esmx": "./dist/cli.mjs"
+        "create-esmx": "./dist/create.mjs"
     },
     "template": "library-node",
     "scripts": {
