+++ packages/create-esmx/src/cli.integration.test.ts
@@ -1,11 +1,12 @@
 import { existsSync } from 'node:fs';
+import { readFileSync } from 'node:fs';
 import { mkdir, mkdtemp, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
-import { createProject } from './index';
+import { cli } from './cli';
+import { getAvailableTemplates } from './template';
 
-// Test utilities
 async function createTempDir(prefix = 'esmx-test-'): Promise<string> {
     return mkdtemp(join(tmpdir(), prefix));
 }
@@ -18,7 +19,88 @@ async function cleanupTempDir(tempDir: string): Promise<void> {
     }
 }
 
-describe('create-esmx integration tests', () => {
+async function verifyProjectStructure(
+    projectPath: string,
+    projectName: string
+): Promise<void> {
+    expect(existsSync(projectPath)).toBe(true);
+    expect(existsSync(join(projectPath, 'src'))).toBe(true);
+
+    const requiredFiles = [
+        'package.json',
+        'tsconfig.json',
+        'README.md',
+        'src/entry.client.ts',
+        'src/entry.node.ts',
+        'src/create-app.ts'
+    ];
+
+    for (const file of requiredFiles) {
+        expect(existsSync(join(projectPath, file))).toBe(true);
+        if (!existsSync(join(projectPath, file))) {
+            throw new Error(`Missing required file: ${file}`);
+        }
+    }
+
+    const packageJson = JSON.parse(
+        readFileSync(join(projectPath, 'package.json'), 'utf-8')
+    );
+
+    const typeCheckCommands = ['vue-tsc --noEmit', 'tsc --noEmit'];
+    const typeGenCommands = [
+        'vue-tsc --declaration --emitDeclarationOnly --noEmit false --outDir dist/src && tsc-alias -p tsconfig.json --outDir dist/src',
+        'tsc --declaration --emitDeclarationOnly --outDir dist/src && tsc-alias -p tsconfig.json --outDir dist/src'
+    ];
+
+    expect(packageJson).toMatchObject({
+        name: projectName,
+        type: 'module',
+        private: true,
+        scripts: {
+            dev: 'esmx dev',
+            build: 'esmx build',
+            preview: 'esmx preview',
+            start: 'NODE_ENV=production node dist/index.mjs'
+        },
+        dependencies: {
+            '@esmx/core': expect.any(String)
+        },
+        devDependencies: {
+            typescript: expect.any(String),
+            '@types/node': expect.any(String),
+            'tsc-alias': expect.any(String)
+        }
+    });
+
+    expect(packageJson.scripts['lint:type']).toBeOneOf(typeCheckCommands);
+    expect(packageJson.scripts['build:type']).toBeOneOf(typeGenCommands);
+
+    const tsconfig = JSON.parse(
+        readFileSync(join(projectPath, 'tsconfig.json'), 'utf-8')
+    );
+
+    expect(tsconfig).toMatchObject({
+        compilerOptions: {
+            module: 'ESNext',
+            moduleResolution: 'node',
+            target: 'ESNext',
+            strict: true,
+            baseUrl: '.',
+            paths: {
+                [`${projectName}/src/*`]: ['./src/*'],
+                [`${projectName}/*`]: ['./*']
+            }
+        },
+        include: ['src'],
+        exclude: ['dist', 'node_modules']
+    });
+
+    const readmeContent = readFileSync(join(projectPath, 'README.md'), 'utf-8');
+    expect(readmeContent.length).toBeGreaterThan(0);
+    expect(readmeContent).toContain(projectName);
+}
+
+describe('create-esmx CLI integration tests', () => {
     let tmpDir: string;
 
     beforeEach(async () => {
@@ -29,58 +111,43 @@ describe('create-esmx integration tests', () => {
         await cleanupTempDir(tmpDir);
     });
 
-    it('should create project with vue2 template', async () => {
-        const projectPath = join(tmpDir, 'test-project');
+    it('should create project with all available templates', async () => {
+        const templates = getAvailableTemplates();
+        expect(templates.length).toBeGreaterThan(0);
 
-        await createProject({
-            argv: ['test-project', '--template', 'vue2'],
-            cwd: tmpDir,
-            userAgent: 'npm/test'
-        });
+        for (const template of templates) {
+            const projectName = `test-${template.folder}`;
+            const projectPath = join(tmpDir, projectName);
 
-        // Verify project directory exists
-        expect(existsSync(projectPath)).toBe(true);
+            await cli({
+                argv: [projectName, '--template', template.folder],
+                cwd: tmpDir,
+                userAgent: 'npm/test'
+            });
 
-        // Verify essential common files exist
-        expect(existsSync(join(projectPath, 'package.json'))).toBe(true);
-        expect(existsSync(join(projectPath, 'tsconfig.json'))).toBe(true);
-        expect(existsSync(join(projectPath, 'README.md'))).toBe(true);
-
-        // Verify src directory exists
-        expect(existsSync(join(projectPath, 'src'))).toBe(true);
-
-        // Verify Esmx common entry files exist
-        expect(existsSync(join(projectPath, 'src/entry.client.ts'))).toBe(true);
-        expect(existsSync(join(projectPath, 'src/entry.node.ts'))).toBe(true);
-        expect(existsSync(join(projectPath, 'src/entry.server.ts'))).toBe(true);
-        expect(existsSync(join(projectPath, 'src/create-app.ts'))).toBe(true);
+            await verifyProjectStructure(projectPath, projectName);
+        }
     });
 
     it('should handle --force parameter correctly', async () => {
         const projectPath = join(tmpDir, 'test-project');
 
-        // Create project first time
-        await createProject({
+        await cli({
             argv: ['test-project', '--template', 'vue2'],
             cwd: tmpDir,
             userAgent: 'npm/test'
         });
 
-        // Verify project exists
         expect(existsSync(join(projectPath, 'package.json'))).toBe(true);
 
-        // Create project again with force flag
-        await createProject({
+        await cli({
             argv: ['test-project', '--template', 'vue2', '--force'],
             cwd: tmpDir,
             userAgent: 'npm/test'
         });
 
-        // Verify project still exists and is valid
         expect(existsSync(join(projectPath, 'package.json'))).toBe(true);
         expect(existsSync(join(projectPath, 'src'))).toBe(true);
-
-        // Verify Esmx common entry files still exist after force overwrite
         expect(existsSync(join(projectPath, 'src/entry.client.ts'))).toBe(true);
         expect(existsSync(join(projectPath, 'src/entry.node.ts'))).toBe(true);
         expect(existsSync(join(projectPath, 'src/entry.server.ts'))).toBe(true);
@@ -88,7 +155,6 @@ describe('create-esmx integration tests', () => {
     });
 
     it('should show help information', async () => {
-        // Mock console.log to capture help output
         const originalLog = console.log;
         const logOutput: string[] = [];
         console.log = (...args: any[]) => {
@@ -96,7 +162,7 @@ describe('create-esmx integration tests', () => {
         };
 
         try {
-            await createProject({
+            await cli({
                 argv: ['--help'],
                 cwd: tmpDir,
                 userAgent: 'npm/test'
@@ -112,7 +178,6 @@ describe('create-esmx integration tests', () => {
     });
 
     it('should show version information', async () => {
-        // Mock console.log to capture version output
         const originalLog = console.log;
         const logOutput: string[] = [];
         console.log = (...args: any[]) => {
@@ -120,7 +185,7 @@ describe('create-esmx integration tests', () => {
         };
 
         try {
-            await createProject({
+            await cli({
                 argv: ['--version'],
                 cwd: tmpDir,
                 userAgent: 'npm/test'
@@ -136,13 +201,12 @@ describe('create-esmx integration tests', () => {
     it('should handle creating directory when target directory does not exist', async () => {
         const projectPath = join(tmpDir, 'non-existent-parent', 'test-project');
 
-        await createProject({
+        await cli({
             argv: ['non-existent-parent/test-project', '--template', 'vue2'],
             cwd: tmpDir,
             userAgent: 'npm/test'
         });
 
-        // Verify project was created in nested directory
         expect(existsSync(projectPath)).toBe(true);
         expect(existsSync(join(projectPath, 'package.json'))).toBe(true);
         expect(existsSync(join(projectPath, 'src'))).toBe(true);
@@ -151,59 +215,50 @@ describe('create-esmx integration tests', () => {
     it('should handle force overwrite for non-empty directory', async () => {
         const projectPath = join(tmpDir, 'test-project');
 
-        // Create directory with some files
         await mkdir(projectPath, { recursive: true });
         await writeFile(
             join(projectPath, 'existing-file.txt'),
             'existing content'
         );
 
-        // Create project with force flag in non-empty directory
-        await createProject({
+        await cli({
             argv: ['test-project', '--template', 'vue2', '--force'],
             cwd: tmpDir,
             userAgent: 'npm/test'
         });
 
-        // Verify project was created successfully
         expect(existsSync(join(projectPath, 'package.json'))).toBe(true);
         expect(existsSync(join(projectPath, 'src'))).toBe(true);
     });
 
     it('should handle force overwrite in current directory', async () => {
-        // Create some files in current directory
         const testFile = join(tmpDir, 'existing-file.txt');
         await writeFile(testFile, 'existing content');
 
-        // Create project in current directory with force flag
-        await createProject({
+        await cli({
             argv: ['.', '--template', 'vue2', '--force'],
             cwd: tmpDir,
             userAgent: 'npm/test'
         });
 
-        // Verify project was created successfully in current directory
         expect(existsSync(join(tmpDir, 'package.json'))).toBe(true);
         expect(existsSync(join(tmpDir, 'src'))).toBe(true);
         expect(existsSync(join(tmpDir, 'src/entry.client.ts'))).toBe(true);
     });
 
     it('should create project in current directory when target is "."', async () => {
-        // Create project in current directory
-        await createProject({
+        await cli({
             argv: ['.', '--template', 'vue2'],
             cwd: tmpDir,
             userAgent: 'npm/test'
         });
 
-        // Verify project was created in current directory
         expect(existsSync(join(tmpDir, 'package.json'))).toBe(true);
         expect(existsSync(join(tmpDir, 'src'))).toBe(true);
         expect(existsSync(join(tmpDir, 'src/entry.client.ts'))).toBe(true);
     });
 
     it('should handle various project name formats', async () => {
-        // Test with different naming styles
         const testCases = [
             'simple-name',
             'nested/project-name',
@@ -213,7 +268,7 @@ describe('create-esmx integration tests', () => {
         for (const projectName of testCases) {
             const projectPath = join(tmpDir, projectName);
 
-            await createProject({
+            await cli({
                 argv: [projectName, '--template', 'vue2'],
                 cwd: tmpDir,
                 userAgent: 'npm/test'
