+++ packages/create-esmx/src/create.ts
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+
+import { cli } from './cli';
+
+cli().catch((error) => {
+    console.error('Error creating project:', error);
+    process.exit(1);
+});
